+++ src/helpers/key.ts
@@ -1,24 +1,61 @@
-import { getServiceTokenData } from '../api';
-import { decryptSymmetric128BitHexKeyUTF8 } from '../utils/crypto';
+import { AuthMode } from "../variables";
 import { 
-    ServiceTokenClientConfig, 
+    getServiceTokenData,
+    getServiceTokenDataKey
+} from '../api';
+import { 
+    decryptSymmetric128BitHexKeyUTF8,
+    decryptAsymmetric
+} from '../utils/crypto';
+import { 
+    ClientConfig,
     WorkspaceConfig
 } from '../types/InfisicalClient';
 
-export const populateClientWorkspaceConfigsHelper = async (clientConfig: ServiceTokenClientConfig): Promise<WorkspaceConfig> => {
-    const serviceTokenData = await getServiceTokenData({
-        apiRequest: clientConfig.apiRequest
-    });
+export const populateClientWorkspaceConfigsHelper = async (clientConfig: ClientConfig): Promise<WorkspaceConfig> => {
+    switch (clientConfig.authMode) {
+        case AuthMode.SERVICE_TOKEN: {
+            const serviceTokenData = await getServiceTokenData({
+                apiRequest: clientConfig.apiRequest
+            });
 
-    const workspaceKey = decryptSymmetric128BitHexKeyUTF8({
-        ciphertext: serviceTokenData.encryptedKey,
-        iv: serviceTokenData.iv,
-        tag: serviceTokenData.tag,
-        key: clientConfig.credentials.serviceTokenKey
-    });
+            const workspaceKey = decryptSymmetric128BitHexKeyUTF8({
+                ciphertext: serviceTokenData.encryptedKey,
+                iv: serviceTokenData.iv,
+                tag: serviceTokenData.tag,
+                key: clientConfig.credentials.serviceTokenKey
+            });
+            
+            return ({
+                workspaceId: serviceTokenData.workspace,
+                workspaceKey
+            });
+        }
+        case AuthMode.SERVICE_TOKEN_V3: {
+            const {
+                key: {
+                    workspace,
+                    encryptedKey,
+                    nonce,
+                    publicKey
+                }
+            } = await getServiceTokenDataKey({
+                apiRequest: clientConfig.apiRequest
+            });
+            
+            const workspaceKey = decryptAsymmetric({
+                ciphertext: encryptedKey,
+                nonce,
+                publicKey,
+                privateKey: clientConfig.credentials.privateKey
+            });
+            
+            return ({
+                workspaceId: workspace,
+                workspaceKey
+            });
+        }
+    }
     
-    return ({
-        workspaceId: serviceTokenData.workspace,
-        workspaceKey
-    });
+    throw Error("Failed to populate workspace config");
 }